refactored file sidebar
