+++ packages/react/src/views/FileMessage/FilePreviewContainer.js
@@ -5,7 +5,7 @@ import { filePreviewContainerStyles as styles } from './Files.styles';
 
 const FilePreviewContainer = ({ file }) => (
   <Box css={styles.previewContainer}>
-    <Avatar url={file.url} fallbackIcon="attachment" size="2.25em" />
+    <Avatar url={file?.url} fallbackIcon="attachment" size="2.25em" />
   </Box>
 );
