--- packages/react/src/views/Files/index.js
@@ -1,1 +1,0 @@
-export { default as Files } from './Files';
